fcode: line wrapping and indentation + one failing test
The failing test is due to a weakness in the printer design, which will
be fixed in the next patch. When objects offer their own printmethods,
they have to call fcode (or another analogon) to get the representations
of their arguments. This way the settings of the printer are not passed
on. A second problem is typical for fcode: the indentation and wrapping
should not be applied to arguments, while this is unavoidable in case of
fcode.